validação, internacionalização e paginação

+++ app/Controllers/Http/EventController.js
@@ -3,45 +3,84 @@
 const Event = use('App/Models/Event')
 
 class EventController {
-  async index ({ request, response, view }) {
-    const events = await Event.query()
-      .with('user')
-      .fetch()
+  async index ({ request, auth }) {
+    try {
+      const { page } = request.get()
 
-    return events
+      const { id } = await auth.getUser()
+      const events = await Event.query()
+        .where('user_id', id)
+        .with('user')
+        .orderBy('date', 'asc')
+        .orderBy('time', 'asc')
+        .paginate(page)
+
+      return events
+    } catch (err) {
+      return err
+    }
   }
 
   async store ({ request, response, auth }) {
-    const data = request.only(['title', 'location', 'dt'])
+    try {
+      const data = request.only(['title', 'location', 'date', 'time'])
+      const event = await Event.create({ ...data, user_id: auth.user.id })
 
-    const event = await Event.create({ ...data, user_id: auth.user.id })
-
-    return event
+      return event
+    } catch (err) {
+      return err
+    }
   }
 
-  async show ({ params }) {
-    const event = await Event.findOrFail(params.id)
+  async show ({ response, params, auth }) {
+    try {
+      const { id } = await auth.getUser()
 
-    await event.load('user')
+      const event = await Event.findOrFail(params.id)
 
-    return event
+      if (event.user_id !== id) {
+        return "You cannot see someone else's events."
+      }
+      await event.load('user')
+      return event
+    } catch (err) {
+      return err
+    }
   }
 
-  async update ({ params, request }) {
-    const event = await Event.findOrFail(params.id)
-    const data = request.only(['title', 'location', 'dt'])
+  async update ({ response, params, request, auth }) {
+    try {
+      const { id } = await auth.getUser()
 
-    event.merge(data)
+      const event = await Event.findOrFail(params.id)
+      const data = request.only(['title', 'location', 'date', 'time'])
 
-    await event.save()
+      if (event.user_id !== id) {
+        return "You cannot edit someone else's events."
+      }
+      event.merge(data)
 
-    return event
+      await event.save()
+
+      return event
+    } catch (err) {
+      return err
+    }
   }
 
-  async destroy ({ params }) {
-    const event = await Event.findOrFail(params.id)
+  async destroy ({ params, auth }) {
+    try {
+      const { id } = await auth.getUser()
 
-    await event.delete()
+      const event = await Event.findOrFail(params.id)
+
+      if (event.user_id !== id) {
+        return "You cannot delete someone else's events."
+      }
+      await event.delete()
+    } catch (err) {
+      return err
+    }
   }
 }
